requirements.txt
mysql-connector-python
tkPDFViewer

